# Speed up flood fill using priority queue and grids