--- a/src/map.cpp
+++ b/src/map.cpp
@@ -8,6 +8,7 @@
 #include <cstring>
 #include <iterator>
 #include <limits>
+#include <queue>
 #include <sstream>
 #include <type_traits>
 #include <unordered_map>
@@ -6210,40 +6211,108 @@ std::vector<tripoint> map::find_clear_path( const tripoint &source,
 void map::reachable_flood_steps( std::vector<tripoint> &reachable_pts, const tripoint &f,
                                  int range, const int cost_min, const int cost_max ) const
 {
-    // the starting spot for the search should definitely be included
-    reachable_pts.push_back( f );
-    std::vector < std::pair<tripoint, int>> f_map = { {f, 0} };
-    // iterate out to the maximum number of steps
-    for( unsigned int r = 1; r < range; r++ ) {
-        int f_map_size = f_map.size();
-        for( unsigned int k = 0; k < f_map_size; k++ ) {
-            for( int i = -1; i <= 1; i++ ) {
-                for( int j = -1; j <= 1; j++ ) {
-                    tripoint tp = { f_map[k].first.x + i, f_map[k].first.y + j, f.z };
-                    const int cost = this->move_cost( { tp.x, tp.y, tp.z } );
-                    // rejection conditions
-                    if( ( tp.x == f_map[k].first.x && tp.y == f_map[k].first.y ) ||
-                        cost < cost_min || cost > cost_max || !has_floor_or_support( tp ) ) {
-                        continue;
-                    } else {
-                        // now make sure we haven't already stepped on this spot
-                        bool np = true;
-                        for( unsigned int l = 0; l < f_map.size(); l++ ) {
-                            if( tp == f_map[l].first ) {
-                                np = false;
-                                break;
-                            }
-                        }
-                        if( !np ) {
-                            continue;
-                        } else {
-                            // add this point to our lists of reachables
-                            f_map.push_back( { tp, r } );
-                            reachable_pts.push_back( tp );
+    struct pq_item {
+        int dist;
+        int ndx;
+    };
+    struct pq_item_comp {
+        bool operator()( const pq_item &left, const pq_item &right ) {
+            return left.dist > right.dist;
+        }
+    };
+    using PQ_type = std::priority_queue< pq_item, std::vector<pq_item>, pq_item_comp>;
+
+    // temp buffer for grid
+    const int grid_dim = range * 2 + 1;
+    std::vector< int > t_grid( grid_dim * grid_dim, -1 ); // init to -1 as "not visited yet"
+    const tripoint origin_offset = {range, range, 0};
+    const int initial_visit_distance = range * range; // Large unreachable value
+
+    // Fill positions that are visitable with initial_visit_distance
+    for( const tripoint &p : points_in_radius( f, range ) ) {
+        const tripoint tp = { p.x, p.y, f.z };
+        const int tp_cost = move_cost( tp );
+        // rejection conditions
+        if( tp_cost < cost_min || tp_cost > cost_max || !has_floor_or_support( tp ) ) {
+            continue;
+        }
+        // set initial cost for grid point
+        tripoint origin_relative = tp - f;
+        origin_relative += origin_offset;
+        int ndx = origin_relative.x + origin_relative.y * grid_dim;
+        t_grid[ ndx ] = initial_visit_distance;
+    }
+    
+    auto gen_neighbors = []( const pq_item &elem, int grid_dim, pq_item *neighbors ) {
+        // Up to 8 neighbors
+        int new_cost = elem.dist + 1;
+        int ox[8] = {
+            -1, 0, 1,
+            -1,    1,
+            -1, 0, 1
+        };
+        int oy[8] = {
+            -1, -1, -1,
+             0,      0,
+             1,  1,  1
+        };
+
+        int ex = elem.ndx % grid_dim;
+        int ey = elem.ndx / grid_dim;
+        for( int i = 0; i < 8; ++i ) {
+            int nx = ex + ox[i];
+            int ny = ey + oy[i];
+
+            int ndx = nx + ny * grid_dim;
+            neighbors[i] = { new_cost, ndx };
+        }
+    };
+
+    PQ_type pq( pq_item_comp{} );
+    pq_item first_item{ 0, range + range * grid_dim };
+    pq.push( first_item );
+    pq_item neighbor_elems[8];
+
+    while( !pq.empty() ) {
+        const pq_item item = pq.top();
+        pq.pop();
+
+        if( t_grid[ item.ndx ] == initial_visit_distance ) {
+            t_grid[ item.ndx ] = item.dist;
+            if( item.dist + 1 < range ) {
+                gen_neighbors( item, grid_dim, neighbor_elems );
+                for( int i = 0; i < 8; ++i ) {
+                    pq.push( neighbor_elems[i] );
+                }
+            }
+        }
+    }
+    std::vector<char> o_grid( grid_dim * grid_dim, 0 );
+    for( int y = 0, ndx = 0; y < grid_dim; ++y ) {
+        for( int x = 0; x < grid_dim; ++x, ++ndx ) {
+            if( t_grid[ ndx ] != -1 && t_grid[ ndx ] < initial_visit_distance ) {
+                // set self and neighbors to 1
+                for( int dy = -1; dy <= 1; ++dy ) {
+                    for( int dx = -1; dx <= 1; ++dx ) {
+                        int tx = dx + x;
+                        int ty = dy + y;
+
+                        if( tx >= 0 && tx < grid_dim && ty >= 0 && ty < grid_dim ) {
+                            o_grid[ tx + ty * grid_dim ] = 1;
                         }
                     }
                 }
-
+            }
+        }
+    }
+    // Remove origin from output set
+    o_grid[ range + range * grid_dim ] = 0;
+    // Now go over again to pull out all of the reachable points
+    for( int y = 0, ndx = 0; y < grid_dim; ++y ) {
+        for( int x = 0; x < grid_dim; ++x, ++ndx ) {
+            if( o_grid[ ndx ] ) {
+                tripoint t = f - origin_offset + tripoint{ x, y, 0 };
+                reachable_pts.push_back( t );
             }
         }
     }
@@ -6252,7 +6321,7 @@ void map::reachable_flood_steps( std::vector<tripoint> &reachable_pts, const tri
 bool map::check_reachables( const std::vector<tripoint> &reachable_pts, const tripoint &f ) const
 {
     const int s = reachable_pts.size();
-    for( unsigned int i = 0; i < s; i++ ) {
+    for( int i = 0; i < s; i++ ) {
         // if we can physically walk to a spot, then we can grab things from it's neighboring tiles too
         if( reachable_pts[i] == f || ( abs( f.x - reachable_pts[i].x ) <= 1 &&
                                        abs( f.y - reachable_pts[i].y ) <= 1 ) ) {