#### Summary
```SUMMARY: Performance "Speed up flood fill in inventory::form_from_map"```

#### Purpose of change
In the extremely large range case (eg R=20, R=50) iterating over tripoints to see if a candidate point has been picked already is slow.

#### Describe the solution
Precondition data to fit in a grid, then run a breadth-first undirected search to find all connected points to the origin. Then expand each of these connected points to include their adjacent points to remove the need to check inside of the loop.

Only need to run the flood fill if `clear_path == true`. If it is `false` we want everything in the range and can just use the points in radius. Conditioning the data earlier hoists loop invariants, removes branching, and speeds things up.

#### Describe alternatives you've considered
Leave it alone: the only time we need to actually do the flood fill, in the current state of the code, is at Range = 6 the current solution is probably fast enough. 

#### Additional context
```
PROFILE RESULTS [R=50] (Current Implementation)
26188758955 ns : inventory::form_from_map
26149242731 ns : inventory::form_from_map


PROFILE RESULTS [R=50] (PR Implementation)
   29070601 ns : inventory::form_from_map
   27311804 ns : inventory::form_from_map
   34487237 ns : inventory::form_from_map
```